requirements.txt
pymongo==4.6.1
dnspython==2.4.2
beautifulsoup4
requests
google-api-python-client
google-auth-oauthlib
python-dateutil 